chore: convert docs to TS
