+++ docs/package.json
@@ -13,8 +13,8 @@
   },
   "dependencies": {
     "color": "^3.1.2",
-    "component-docs": "^0.23.0",
-    "linaria": "^1.2.4",
+    "component-docs": "0.24.0",
+    "linaria": "^2.1.1",
     "react-native-vector-icons": "^6.6.0"
   },
   "devDependencies": {
@@ -22,12 +22,9 @@
     "@babel/node": "^7.6.3",
     "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/preset-env": "^7.8.4",
-    "@babel/preset-flow": "^7.8.3",
     "@babel/preset-react": "^7.8.3",
     "@babel/preset-typescript": "^7.8.3",
-    "del-cli": "^3.0.0",
     "file-loader": "^4.2.0",
-    "metro-react-native-babel-preset": "^0.56.0",
     "url-loader": "^4.1.0"
   }
 }
