+++ docs/pages/src/components/IphoneIcon.tsx
@@ -1,9 +1,7 @@
-/* @flow */
-
 import React from 'react';
 
 type Props = {
-  color: string,
+  color: string;
 };
 
 const IphoneIcon = ({ color }: Props) => (
