+++ scripts/generate-ts-tests.js
@@ -1,6 +1,6 @@
-const glob = require('glob');
 const { readFileSync, writeFileSync, mkdirSync } = require('fs');
 const path = require('path');
+const glob = require('glob');
 const rimraf = require('rimraf');
 
 const SOURCE_FILES_PATTERN = './src/**/*.js';
